add mui and fix right-panel
