+++ package.json
@@ -10,11 +10,13 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@emotion/react": "^11.13.3",
+    "@emotion/styled": "^11.13.0",
+    "@mui/material": "^6.1.3",
     "@types/uuid": "^10.0.0",
     "appwrite": "^14.0.1",
     "clsx": "^2.1.1",
     "framer-motion": "^11.11.4",
-    "hamburger-react": "^2.5.1",
     "lucide-react": "^0.447.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
