+++ src/pages/admin/dashboard/components/main-panel.jsx
@@ -67,7 +67,3 @@ const MainPanel = () => {
 };
 
 export default MainPanel;
-
-
-
-
