+++ src/pages/admin/dashboard/dashboard.jsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { useUserContext } from "../../../providers/UserProvider";
 import { useParams } from "react-router-dom";
-import RightPanel from "./components/right-panel";
 import TopBar from "../../../components/topbar";
 import MainPanel from "./components/main-panel";
 
@@ -21,9 +20,8 @@ const DashboardPage = () => {
   return (
     <div className="w-full h-[100dvh] flex flex-col">
       <TopBar category={"ORG"} />
-      <div className="relative h-full flex w-full">
+      <div className="relative h-full flex w-full justify-center items-start">
         <MainPanel />
-        <RightPanel />
       </div>
     </div>
   );
